[KAN-37, KAN-39] 갓생지수, 참새/올빼미족  컴포넌트 생성 완료
