--- app/dashboard/tmp-GodLifeIndex/page.tsx
@@ -1,13 +1,0 @@
-// app/dashboard/tmp-GodLifeIndex/page.tsx
-'use client';
-
-import { GodLifIndex } from "@/components/component/GodLifeIndex; 
-
-export default function GodLifeIndexPage() {
-  return (
-    <main className="flex min-h-screen flex-col items-center justify-center">
-      {GodLifIndex()}
-      
-    </main>
-  );
-}
