+++ app/layout.tsx
@@ -34,7 +34,8 @@ export const metadata: Metadata = {
 };
 
 if (!process.env.NEXT_PUBLIC_GOOGLE_CLIENT_ID) {
-  console.log('Missing Google Client ID');
+  throw new Error('Missing Google Client ID'); 
+  // console.log('Missing Google Client ID'); // 로컬에서 작업하기 위해서는 error throw가 아니라 console.log 처리
 }
 
 const clientId: string = process.env.NEXT_PUBLIC_GOOGLE_CLIENT_ID;
