+++ package.json
@@ -24,7 +24,7 @@
     "date-fns": "^3.6.0",
     "jwt-decode": "^4.0.0",
     "lucide-react": "^0.439.0",
-    "next": "14.2.8",
+    "next": "^14.2.8",
     "react": "^18",
     "react-day-picker": "^8.10.1",
     "react-dom": "^18",
